AM-1087: Optimising DB queries.
