+++ src/main/java/uk/gov/hmcts/reform/roleassignment/util/PersistenceUtil.java
@@ -19,6 +19,8 @@ import uk.gov.hmcts.reform.roleassignment.domain.model.enums.Status;
 
 import java.util.Arrays;
 import java.util.Collections;
+import java.util.Objects;
+import java.util.UUID;
 import java.util.stream.Collectors;
 
 @Service
@@ -149,5 +151,22 @@ public class PersistenceUtil {
                                 ? Arrays.asList(roleAssignmentEntity.getAuthorisations().split(";")) :
                                 Collections.emptyList())
             .build();
+    }
+
+    public HistoryEntity prepareHistoryEntityForPersistance(RoleAssignment roleAssignment, Request request) {
+        UUID roleAssignmentId = roleAssignment.getId();
+        UUID requestId = request.getId();
+
+        RequestEntity requestEntity = convertRequestToEntity(request);
+        if (requestId != null) {
+            requestEntity.setId(requestId);
+        }
+
+        HistoryEntity historyEntity = convertRoleAssignmentToHistoryEntity(
+            roleAssignment,
+            requestEntity
+        );
+        historyEntity.setId(Objects.requireNonNullElseGet(roleAssignmentId, UUID::randomUUID));
+        return historyEntity;
     }
 }
