+++ src/test/java/uk/gov/hmcts/reform/roleassignment/domain/service/createroles/CreateRoleAssignmentServiceTest.java
@@ -105,7 +105,7 @@ class CreateRoleAssignmentServiceTest {
         when(parseRequestService.parseRequest(any(AssignmentRequest.class), any(RequestType.class))).thenReturn(
             existingAssignmentRequest);
         when(persistenceService.persistRequest(any(Request.class))).thenReturn(requestEntity);
-        when(persistenceService.persistHistory(
+        when(persistenceUtil.prepareHistoryEntityForPersistance(
             any(RoleAssignment.class),
             any(Request.class)
         )).thenReturn(historyEntity);
@@ -122,8 +122,8 @@ class CreateRoleAssignmentServiceTest {
             .deleteRoleAssignment(any(RoleAssignment.class));
         verify(persistenceService, times(2))
             .persistActorCache(any(RoleAssignment.class));
-        verify(persistenceService, times(2))
-            .persistHistory(any(RoleAssignment.class), any(Request.class));
+        verify(persistenceUtil, times(2))
+            .prepareHistoryEntityForPersistance(any(RoleAssignment.class), any(Request.class));
 
     }
 
@@ -156,7 +156,7 @@ class CreateRoleAssignmentServiceTest {
         when(parseRequestService.parseRequest(any(AssignmentRequest.class), any(RequestType.class))).thenReturn(
             existingAssignmentRequest);
         when(persistenceService.persistRequest(any(Request.class))).thenReturn(requestEntity);
-        when(persistenceService.persistHistory(
+        when(persistenceUtil.prepareHistoryEntityForPersistance(
             any(RoleAssignment.class),
             any(Request.class)
         )).thenReturn(historyEntity);
@@ -169,8 +169,8 @@ class CreateRoleAssignmentServiceTest {
         //assertion
         verify(persistenceService, times(5))
             .updateRequest(any(RequestEntity.class));
-        verify(persistenceService, times(8))
-            .persistHistory(any(RoleAssignment.class), any(Request.class));
+        verify(persistenceUtil, times(8))
+            .prepareHistoryEntityForPersistance(any(RoleAssignment.class), any(Request.class));
         verify(validationModelService, times(1))
             .validateRequest(any(AssignmentRequest.class));
 
@@ -205,7 +205,7 @@ class CreateRoleAssignmentServiceTest {
         when(parseRequestService.parseRequest(any(AssignmentRequest.class), any(RequestType.class))).thenReturn(
             existingAssignmentRequest);
         when(persistenceService.persistRequest(any(Request.class))).thenReturn(requestEntity);
-        when(persistenceService.persistHistory(
+        when(persistenceUtil.prepareHistoryEntityForPersistance(
             any(RoleAssignment.class),
             any(Request.class)
         )).thenReturn(historyEntity);
@@ -223,8 +223,8 @@ class CreateRoleAssignmentServiceTest {
             .deleteRoleAssignment(any(RoleAssignment.class));
         verify(persistenceService, times(2))
             .persistActorCache(any(RoleAssignment.class));
-        verify(persistenceService, times(2))
-            .persistHistory(any(RoleAssignment.class), any(Request.class));
+        verify(persistenceUtil, times(2))
+            .prepareHistoryEntityForPersistance(any(RoleAssignment.class), any(Request.class));
 
     }
 
@@ -284,7 +284,7 @@ class CreateRoleAssignmentServiceTest {
 
         when(persistenceUtil.convertHistoryEntityToRoleAssignment(any(HistoryEntity.class)))
             .thenReturn(existingAssignmentRequest.getRequestedRoles().iterator().next());
-        when(persistenceService.persistHistory(
+        when(persistenceUtil.prepareHistoryEntityForPersistance(
             any(RoleAssignment.class),
             any(Request.class)
         )).thenReturn(historyEntity);
@@ -299,8 +299,8 @@ class CreateRoleAssignmentServiceTest {
 
         verify(persistenceService, times(1))
             .persistActorCache(any(RoleAssignment.class));
-        verify(persistenceService, times(2))
-            .persistHistory(any(RoleAssignment.class), any(Request.class));
+        verify(persistenceUtil, times(2))
+            .prepareHistoryEntityForPersistance(any(RoleAssignment.class), any(Request.class));
         verify(persistenceService, times(1))
             .persistRoleAssignment(any(RoleAssignment.class));
 
@@ -332,7 +332,7 @@ class CreateRoleAssignmentServiceTest {
 
         when(persistenceUtil.convertHistoryEntityToRoleAssignment(any(HistoryEntity.class)))
             .thenReturn(existingAssignmentRequest.getRequestedRoles().iterator().next());
-        when(persistenceService.persistHistory(
+        when(persistenceUtil.prepareHistoryEntityForPersistance(
             any(RoleAssignment.class),
             any(Request.class)
         )).thenReturn(historyEntity);
@@ -346,8 +346,8 @@ class CreateRoleAssignmentServiceTest {
 
         verify(persistenceService, times(0))
             .persistActorCache(any(RoleAssignment.class));
-        verify(persistenceService, times(0))
-            .persistHistory(any(RoleAssignment.class), any(Request.class));
+        verify(persistenceUtil, times(0))
+            .prepareHistoryEntityForPersistance(any(RoleAssignment.class), any(Request.class));
         verify(persistenceService, times(0))
             .persistRoleAssignment(any(RoleAssignment.class));
