+++ src/test/java/uk/gov/hmcts/reform/roleassignment/util/PersistenceUtilTest.java
@@ -74,4 +74,67 @@ class PersistenceUtilTest {
         assertNotNull(persistenceUtil.convertEntityToRoleAssignment(
             entity));
     }
+
+    /*@Test
+    void persistHistory() throws IOException {
+        AssignmentRequest assignmentRequest = TestDataBuilder
+            .buildAssignmentRequest(Status.CREATED, Status.LIVE, false);
+        RequestEntity requestEntity = TestDataBuilder.buildRequestEntity(assignmentRequest.getRequest());
+        HistoryEntity historyEntity = TestDataBuilder.buildHistoryIntoEntity(
+            assignmentRequest.getRequestedRoles().iterator().next(), requestEntity);
+
+        when(persistenceUtil.convertRequestToEntity(assignmentRequest.getRequest())).thenReturn(requestEntity);
+        when(persistenceUtil.convertRoleAssignmentToHistoryEntity(
+            assignmentRequest.getRequestedRoles().iterator().next(), requestEntity)).thenReturn(historyEntity);
+
+        HistoryEntity historyEntityResult = sut.prepareHistoryEntityForPersistance(
+            assignmentRequest.getRequestedRoles().iterator().next(), assignmentRequest.getRequest());
+
+        assertNotNull(historyEntityResult);
+        assertNotNull(assignmentRequest.getRequest().getId());
+        assertNotNull(historyEntityResult.getRequestEntity().getId());
+
+        assertEquals(assignmentRequest.getRequest().getId(), historyEntityResult.getRequestEntity().getId());
+        assertEquals(assignmentRequest.getRequestedRoles().iterator().next().getId(), historyEntityResult.getId());
+        for (RoleAssignment requestedRole : assignmentRequest.getRequestedRoles()) {
+            assertEquals(requestedRole.getId(), historyEntityResult.getId());
+        }
+
+        verify(persistenceUtil, times(1)).convertRequestToEntity(any(Request.class));
+        verify(persistenceUtil, times(1)).convertRoleAssignmentToHistoryEntity(
+            any(RoleAssignment.class), any(RequestEntity.class));
+        verify(entityManager, times(1)).persist(any(HistoryEntity.class));
+    }
+
+    @Test
+    void persistHistory_NullRequestId() throws IOException {
+        AssignmentRequest assignmentRequest = TestDataBuilder
+            .buildAssignmentRequest(Status.CREATED, Status.LIVE, false);
+        assignmentRequest.getRequest().setId(null);
+        assignmentRequest.getRequestedRoles().iterator().next().setId(null);
+        RequestEntity requestEntity = TestDataBuilder.buildRequestEntity(assignmentRequest.getRequest());
+        HistoryEntity historyEntity = TestDataBuilder.buildHistoryIntoEntity(
+            assignmentRequest.getRequestedRoles().iterator().next(), requestEntity);
+
+        when(persistenceUtil.convertRequestToEntity(assignmentRequest.getRequest())).thenReturn(requestEntity);
+        when(persistenceUtil.convertRoleAssignmentToHistoryEntity(any(), any())).thenReturn(historyEntity);
+        when(historyRepository.save(historyEntity)).thenReturn(historyEntity);
+
+        HistoryEntity historyEntityResult = sut.prepareHistoryEntityForPersistance(
+            assignmentRequest.getRequestedRoles().iterator().next(), assignmentRequest.getRequest());
+
+        assertNotNull(historyEntityResult);
+        assertNotNull(historyEntityResult.getId());
+        assertNull(historyEntityResult.getRequestEntity().getId());
+
+        assertEquals(
+            assignmentRequest.getRequestedRoles().iterator().next().getId(),
+            historyEntityResult.getRequestEntity().getId()
+        );
+
+        verify(persistenceUtil, times(1)).convertRequestToEntity(any(Request.class));
+        verify(persistenceUtil, times(1)).convertRoleAssignmentToHistoryEntity(
+            any(RoleAssignment.class), any(RequestEntity.class));
+        verify(entityManager, times(1)).persist(any(HistoryEntity.class));
+    }*/
 }
